+++ decidim-debates/spec/system/explore_debates_spec.rb
@@ -159,6 +159,12 @@ describe "Explore debates", type: :system do
     end
   end
 
+  context "when component is not commentable" do
+    let(:ressources) { create_list(:debate, 3, component: current_component) }
+
+    it_behaves_like "an uncommentable component"
+  end
+
   describe "show" do
     let(:path) do
       decidim_participatory_process_debates.debate_path(